# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Rayna Tours Dubai - API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TourPackage {
  id: number;
  slug: string;
  title: string;
  /** e.g. dubai-city, holiday, desert-safari */
  category: string;
  shortDescription?: string;
  description?: string;
  priceAed: number;
  nights: number;
  days: number;
  imageUrl: string;
  galleryImages?: string[];
  inclusions?: string[];
  exclusions?: string[];
  highlights?: string[];
  isFeatured?: boolean;
  rating?: number;
  reviewCount?: number;
  tags?: string[];
}

export interface ActivityOption {
  name: string;
  priceAed: number;
  description?: string;
  inclusions?: string[];
}

export interface Faq {
  question: string;
  answer: string;
}

export interface Activity {
  id: number;
  slug: string;
  title: string;
  category: string;
  shortDescription?: string;
  description?: string;
  priceAed: number;
  duration: string;
  imageUrl: string;
  galleryImages?: string[];
  inclusions?: string[];
  options?: ActivityOption[];
  faqs?: Faq[];
  relatedActivitySlugs?: string[];
  rating?: number;
  reviewCount?: number;
}

export interface EnquiryInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  activityOrPackage: string;
  travelDate: string;
  /** @minimum 1 */
  guests: number;
  /** @nullable */
  message?: string | null;
}

export interface Enquiry {
  id: number;
  name: string;
  email: string;
  phone: string;
  activityOrPackage: string;
  travelDate: string;
  guests: number;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}

export interface Testimonial {
  id: number;
  name: string;
  location: string;
  rating: number;
  review: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export type ListPackagesParams = {
category?: string;
featured?: boolean;
};