# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Rayna Tours Dubai - API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityOption } from './activityOption';
import type { Faq } from './faq';

export interface Activity {
  id: number;
  slug: string;
  title: string;
  category: string;
  shortDescription?: string;
  description?: string;
  priceAed: number;
  duration: string;
  imageUrl: string;
  galleryImages?: string[];
  inclusions?: string[];
  options?: ActivityOption[];
  faqs?: Faq[];
  relatedActivitySlugs?: string[];
  rating?: number;
  reviewCount?: number;
}