# lib/api-zod/src/generated/types/activityOption.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Rayna Tours Dubai - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityOption {
  name: string;
  priceAed: number;
  description?: string;
  inclusions?: string[];
}