# lib/api-zod/src/generated/types/enquiry.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Rayna Tours Dubai - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Enquiry {
  id: number;
  name: string;
  email: string;
  phone: string;
  activityOrPackage: string;
  travelDate: string;
  guests: number;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}