# lib/api-zod/src/generated/types/enquiryInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Rayna Tours Dubai - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnquiryInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  activityOrPackage: string;
  travelDate: string;
  /** @minimum 1 */
  guests: number;
  /** @nullable */
  message?: string | null;
}