# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Rayna Tours Dubai - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  name: string;
  location: string;
  rating: number;
  review: string;
  /** @nullable */
  avatarUrl?: string | null;
}