# lib/api-zod/src/generated/types/tourPackage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Rayna Tours Dubai - API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TourPackage {
  id: number;
  slug: string;
  title: string;
  /** e.g. dubai-city, holiday, desert-safari */
  category: string;
  shortDescription?: string;
  description?: string;
  priceAed: number;
  nights: number;
  days: number;
  imageUrl: string;
  galleryImages?: string[];
  inclusions?: string[];
  exclusions?: string[];
  highlights?: string[];
  isFeatured?: boolean;
  rating?: number;
  reviewCount?: number;
  tags?: string[];
}